% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelToListNA.R
\name{ExcelToListNA}
\alias{ExcelToListNA}
\title{Convert a xlsx file to a list of dataframes retain NAs}
\usage{
ListToExcel(Data = ksearlist, filename = "~/ksear_output.xlsx")
}
\arguments{
\item{exfile}{name of excel focument to put in}
}
\description{
Convert a xlsx file to a list of dataframes retain NAs
}
