% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guava_PDAnalysis.R
\name{Guava_PDAnalysis}
\alias{Guava_PDAnalysis}
\title{Guava PharmacoDynamic Analysis}
\usage{
Guava_PDAnalysis(GuavaSmoothy_object)
}
\arguments{
\item{x}{Normalised GuavaSmoothy object}
}
\value{
Returns a data.frame containing "AUC_Viability", "AUC_CellNo", "AUC_death", "IC50_Viability", "IC50_CellNo", "IC50_Death", "Drug" and "Cell.Line".
}
\description{
Calculated IC50 and AUC values for all metrics for all drug and cell line combinations
}
