% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplots.R
\name{guava_scatter_plot}
\alias{guava_scatter_plot}
\title{Guava Scatterplot}
\usage{
guava_scatter_plot(
  x,
  combo = T,
  drugs = NULL,
  cell_lines = NULL,
  .ncols = 2,
  line.size = 2,
  point.size = 3,
  .rel.widths = c(0.6, 0.2),
  errorbar.size = 0.25
)
}
\arguments{
\item{x}{guava_data}

\item{combo}{if TRUE combinations will be converted for graphing}

\item{drugs}{list of drugs to analyse if NULL all will be used}

\item{cell_lines}{list of cell lines to analyse if NULL all will be used}

\item{.ncols}{number of columns used in output}

\item{line.size}{size of line used}

\item{point.size}{size of datapoints used}

\item{.rel.widths}{relative width ratios between scatterplots and legened}

\item{errorbar.size}{size of errorbars}
}
\description{
Guava Scatterplot
}
