% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guava_synergy.R
\name{Guava_synergy}
\alias{Guava_synergy}
\title{Calculate Guava Synergy Scores}
\usage{
Guava_synergy(x, combinations = NULL, combo_sep = "&", as.percentage = T)
}
\arguments{
\item{x}{Normalise GuavaSmoothy Object}

\item{combinations}{list of drug combinations for synergy calculation. If NULL all combination treatments in data will be used.}

\item{combo_sep}{Seperater used between drug concentrations and names default is "&"}
}
\description{
Calculate Guava Synergy Scores
}
