% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauva_synergy3.R
\name{Guava_synergy3}
\alias{Guava_synergy3}
\title{Calculate Guava Synergy Scores}
\usage{
Guava_synergy3(x, combinations = NULL, combo_sep = "&", as.percentage = T)
}
\arguments{
\item{x}{Normalise GuavaSmoothy Object}

\item{combinations}{list of drug combinations for synergy calculation. If NULL all combination treatments in data will be used.}

\item{combo_sep}{Separator used between drug concentrations and names default is "&"}

\item{as.percentage}{if true viability, cell fold change and cell death values will be treated as percentages}
}
\description{
Calculate Guava Synergy Scores
}
