% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guava_smoothy.R
\name{GuavaSmoothy}
\alias{GuavaSmoothy}
\title{Import values from VIA files}
\usage{
GuavaSmoothy("assay_file.VIA")
}
\arguments{
\item{label}{labels generated from import labels to be used}

\item{filenames}{list of filenames you want to be labeled by the template}

\item{easyfit}{if true values from easyfit will be used}
}
\description{
Import values from VIA files
}
