% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guava_smoothy_dir.R
\name{GuavaSmoothy_dir}
\alias{GuavaSmoothy_dir}
\title{Import values from VIA files}
\usage{
GuavaSmoothy("assay_file.VIA")
}
\arguments{
\item{easyfit}{if true values from easyfit will be used}

\item{filename}{path and filename for target file}

\item{label}{labels generated from import labels to be used}
}
\description{
Import values from VIA files
}
