% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.fcs}
\alias{import.fcs}
\title{Import FCS data from FCS file (*not recommended*)}
\usage{
import.fcs(files_folder)
}
\arguments{
\item{files_folder}{folder containing FCS files}
}
\value{
matrix containing imported FCS data from FCS files
}
\description{
Import FCS data from FCS file (*not recommended*)
}
