% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.fcs.via}
\alias{import.fcs.via}
\title{Import FCS data from VIA file}
\usage{
import.fcs.via(files_folder, .cores = 1)
}
\arguments{
\item{files_folder}{folder containing .VIA.FCS file}

\item{cores}{Number of cores to use for import}
}
\value{
Returns a dataframe contanining all data from the 96 well plate
}
\description{
Import FCS data from VIA file
}
