% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListToExcel.R
\name{ListToExcel}
\alias{ListToExcel}
\title{Convert a list of dataframes/matrices into an excelfile}
\usage{
ListToExcel(Data = ksearlist, filename = "~/ksear_output.xlsx")
}
\arguments{
\item{list}{List of DataFrames to be converted into an xlsx document(names(list) will be used as sheet names)}

\item{filename}{path and filename ending with ".xlsx" for the file to be made to}

\item{rowNames}{if TRUE rownames will be saved in xlsx file}
}
\description{
Convert a list of dataframes/matrices into an excelfile
}
