% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{make.kmeans.clusters}
\alias{make.kmeans.clusters}
\title{Make Kmeans Clusters}
\usage{
make.kmeans.clusters(
  x,
  my.seed = 123,
  .iter.max = 1000,
  .nstart = 50,
  .centers = 2,
  my_minpts = NULL
)
}
\arguments{
\item{x}{Imported FCS data}

\item{my.seed}{seed to use for Kmeans clustering}

\item{.iter.max}{maximum number of kmeas interations}

\item{.nstart}{kmeans nstart}

\item{.centers}{number of centers to use for analysis}

\item{my_minpts}{minimum number of points needed for kmeans to identify a cluster}
}
\value{
return list with raw data and labelled outputs
}
\description{
Make Kmeans Clusters
}
