% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_guava.R
\name{normalise_guave}
\alias{normalise_guave}
\alias{normalise_guava}
\title{Import values from VIA files}
\usage{
GuavaSmoothy("assay_file.VIA")
}
\arguments{
\item{x}{path and filename for target file}
}
\description{
Import values from VIA files
}
