% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Plate_layout_script.R
\name{Read_Plate_layout}
\alias{Read_Plate_layout}
\title{Read Plate Layout file for Guava}
\usage{
Read_Plate_layout("platelayout.xlsx", colwise = TRUE, concentration = FALSE)
}
\arguments{
\item{filename}{Path of the plate layout file}

\item{conc}{If TRUE concentrations will be imported}
}
\description{
Read Plate Layout file for Guava
}
