% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplots.R
\name{Scatter.plot.wizard}
\alias{Scatter.plot.wizard}
\title{Scatter Plot Wizard}
\usage{
Scatter.plot.wizard(x, .combo = F)
}
\arguments{
\item{x}{guava data}

\item{.combo}{if TRUE plots will be treated as combination treatments}
}
\description{
Scatter Plot Wizard
}
